requirements.txt
# Core
python-dotenv==1.0.1
httpx==0.27.0
tenacity==8.2.3

# Data & Analysis
pandas==2.2.1
numpy==1.26.4

# Visualization
matplotlib==3.8.3
plotly==5.19.0
seaborn==0.13.2

# Report Generation
reportlab==4.1.0

# Dashboard (bonus)
streamlit==1.31.1

# Utilities
rich==13.7.0           # Beautiful CLI logging
pydantic==2.6.1        # Data validation & schemas
click==8.1.7           # CLI interface

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
